requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
click==8.2.1
    # via
    #   arrowport (pyproject.toml)
    #   uvicorn
duckdb==1.3.0
    # via arrowport (pyproject.toml)
fastapi==0.115.12
    # via arrowport (pyproject.toml)
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
lz4==4.4.4
    # via arrowport (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
prometheus-client==0.22.0
    # via arrowport (pyproject.toml)
pyarrow==20.0.0
    # via arrowport (pyproject.toml)
pydantic==2.11.5
    # via
    #   arrowport (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via arrowport (pyproject.toml)
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via pydantic-settings
python-multipart==0.0.20
    # via arrowport (pyproject.toml)
pyyaml==6.0.2
    # via arrowport (pyproject.toml)
rich==14.0.0
    # via arrowport (pyproject.toml)
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
structlog==25.3.0
    # via arrowport (pyproject.toml)
typing-extensions==4.13.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.2
    # via arrowport (pyproject.toml)
zstandard==0.23.0
    # via arrowport (pyproject.toml)
